% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_change}
\alias{describe_change}
\title{Describe Change}
\usage{
describe_change(
  before,
  after,
  input_unit,
  output_unit,
  template = "{direction} {change} {unit}",
  direction_phrases = c(increase = "increased by", decrease = "decreased by", none =
    "remained unchanged"),
  include_values = FALSE
)
}
\arguments{
\item{before}{The initial value to compare.}

\item{after}{The value to compare to the initial value.}

\item{input_unit}{One of "number", "percent", or "ratio".}

\item{output_unit}{One of "number", "percent", "times", or "points". "points"
may only be used when \code{input_unit} is "percent" or "ratio".}

\item{template}{A \code{{glue}} template for the string returned when there is a change.}

\item{direction_phrases}{A named vector with three items: "increase", "decrease", and "none",
used to customize either the default or custom template. No one likes to hear
the same thing over and over, after all.}

\item{include_values}{A logical value indicating whether to include \code{before} and \code{after}
in the change description string.}
}
\value{
A string describing the change between two values, optionally including
the values themselves.
}
\description{
Generate descriptive statements about changes between two values.
}
\examples{

# Basic usage with defaults
describe_change(
  before = 1,
  after = 0.5,
  input_unit = "ratio",
  output_unit = "percent",
  template = "{direction} {change} {unit}"
)
#> decreased by 50 percent

# Using different phrasing for changes
describe_change(
  before = 33,
  after = 66,
  input_unit = "percent",
  output_unit = "percent",
  template = "{direction} {change} {unit}",
  direction_phrases = c(
    increase = "rose by",
    decrease = "fell by",
    none = "stagnated"
  )
)
#> rose by 100 percent

# Customizing the template
describe_change(
  before = 10,
  after = 12,
  input_unit = "number",
  output_unit = "number",
  template = "{direction} {change} people"
)
#> increased by 2 people

}
