% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{dir_empty}
\alias{dir_empty}
\title{Directory Empty?}
\usage{
dir_empty(path)
}
\arguments{
\item{path}{A relative or absolute path to the directory to test.}
}
\value{
\code{TRUE} if the directory is empty, otherwise \code{FALSE}.
If the directory doesn't exist, or the string supplied to \code{path} isn't
recognized as a valid path, then the function returns an error.
}
\description{
Takes a path and determines whether it is empty.
}
