% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\docType{data}
\name{ojo_counties}
\alias{ojo_counties}
\title{Oklahoma Counties}
\format{
An object of class \code{character} of length 77.
}
\usage{
ojo_counties
}
\description{
Oklahoma Counties
}
\keyword{datasets}
