% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\name{ojo_list_counties}
\alias{ojo_list_counties}
\title{List Oklahoma Counties}
\usage{
ojo_list_counties(
  case = "lower",
  squish = FALSE,
  suffix = NULL,
  oscn_only = FALSE
)
}
\description{
List Oklahoma Counties
}
