% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\docType{data}
\name{ojo_oscn_counties}
\alias{ojo_oscn_counties}
\title{Oklahoma OSCN reporting counties}
\format{
An object of class \code{character} of length 13.
}
\usage{
ojo_oscn_counties
}
\description{
Oklahoma OSCN reporting counties
}
\keyword{datasets}
