% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojoutils.R
\docType{package}
\name{ojoutils-package}
\alias{ojoutils}
\alias{ojoutils-package}
\title{ojoutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brancen Gregory \email{brancengregory@gmail.com}

}
\keyword{internal}
