% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Convert a decimal to a percentage as a formatted string}
\usage{
percent(
  x,
  digits = 0,
  abs = FALSE,
  drop_trailing_zeros = FALSE,
  multiply_by_100 = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric value to be converted to a percentage}

\item{digits}{Number of decimal places to round to, Default: 0}

\item{abs}{Take the absolute value of x before converting to a percentage, Default: FALSE}

\item{drop_trailing_zeros}{Drop trailing zeros after the decimal point, Default: FALSE}

\item{multiply_by_100}{Multiply x by 100 before converting to a percentage, Default: TRUE}

\item{...}{Additional arguments to be passed to the \code{formatC} function}
}
\value{
A formatted string representing the percentage
}
\description{
User inputs a decimal number and functions converts it to a percentage as a formatted string.
}
\details{
The function uses the \code{round_half_up} function from the janitor package to round the number to the specified number of decimal places. The \code{formatC} function is then used to format the number as a percentage with the specified number of decimal places and trailing zeros.
}
\examples{
x <- 0.1235
x
percent(x)
percent(x, digits = 3)
percent(x, digits = 3, drop_trailing_zeros = FALSE)
}
\seealso{
\code{\link[janitor]{round_half_up}}
\code{\link[base]{formatC}}
}
\author{
Anthony Flores
}
