% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_current_date.R
\name{prepend_current_date}
\alias{prepend_current_date}
\title{Prepend Current Date}
\usage{
prepend_current_date(file_name, append = FALSE)
}
\arguments{
\item{file_name}{A string representing the file name to which the date should be added.}

\item{append}{A logical value indicating whether to append the date instead of prepending it. Default is FALSE.}
}
\value{
A string with the current date added to the file name.
}
\description{
Adds the current date to a file name.
}
\examples{
prepend_current_date("data/output/parties.csv")
prepend_current_date("data/output/parties.csv", append = TRUE)

}
