% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_date.R
\name{stamp_date}
\alias{stamp_date}
\title{Stamp Date}
\usage{
stamp_date(file_name)
}
\arguments{
\item{file_name}{A string representing the file name to which the date should be added.}
}
\value{
A string with the current date added to the file name.
}
\description{
Adds the current date to a file name.
}
\examples{
stamp_date("data/output/parties.csv")

}
