% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{ojo_create_project}
\alias{ojo_create_project}
\title{Create Project}
\usage{
ojo_create_project(name = NULL, dir = ".", private = TRUE)
}
\description{
Creates a new R project with a standard directory structure
}
