% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{ojo_create_project}
\alias{ojo_create_project}
\title{Create Project}
\usage{
ojo_create_project(
  name = NULL,
  description = NULL,
  dir = ".",
  private = TRUE,
  packages = NULL
)
}
\description{
Creates a new R project with a standard directory structure
}
\section{Getting Started}{

To learn more about creating projects, see the vignette:
\code{vignette("project-creation", package = "ojotools")}
}

