% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county.R
\name{ojo_parse_county}
\alias{ojo_parse_county}
\title{Parse Oklahoma Counties}
\usage{
ojo_parse_county(
  county,
  ...,
  case = "lower",
  squish = NULL,
  suffix = NULL,
  counties = ojo_counties,
  .silent = FALSE
)
}
\description{
Parse Oklahoma Counties
}
