% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojo-report-template.R
\name{ojo_use_template}
\alias{ojo_use_template}
\title{Function to use the OJO quarto HTML template}
\usage{
ojo_use_template(template = "openjusticeok/ojo-report-template", directory)
}
\arguments{
\item{template}{The name of the Quarto template to use. This should be a string in the format "username/repository". Default is "openjusticeok/ojo-report-template".}

\item{directory}{The name of the directory where the template should be added. This should be a string representing a path relative to the current working directory.}
}
\value{
This function does not return a value. If the user chooses not to proceed with the operation, the function exits silently.
}
\description{
Wrapper for \verb{quarto use template} command. Defaults to OJO html template.
}
\examples{
\dontrun{
ojo_use_template(template = "openjusticeok/ojo-report-template", directory = "my_directory")
}
}
