% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ojoutils.R
\docType{package}
\name{ojoutils-package}
\alias{ojoutils}
\alias{ojoutils-package}
\title{ojoutils: A Collection of Nifty Functions and Objects for OJO Analysts}
\description{
We find ourselves repeating the same simple tasks or running a series of R commands over and over again. No more!
}
\author{
\strong{Maintainer}: Brancen Gregory \email{brancengregory@gmail.com}

Authors:
\itemize{
  \item Andrew Bell \email{abell@okpolicy.org}
}

}
\keyword{internal}
